backend/requirements.txt
Django==3.2.16
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
django-colorfield==0.11.0
django-filter==23.1
djoser==2.1.0
Pillow==9.3.0